fix(config/workspace): deep-copy RepoOverride.Claude; move git ops to workspace package
RepoOverride.Claude is a pointer field, but copyRepoOverrideMap only did a
shallow struct copy, causing the merged result to share the same *ClaudeOverride
as the original. Downstream mutations to the merged config's hook lists silently
corrupted the original WorkspaceConfig. Replace the shallow copy with
deepCopyRepoOverride, which deep-copies Claude, Env, Files, and SetupDir.

CloneOrSyncOverlay, isValidGitDir, and HeadSHA were in internal/config (a pure
data/validation package) despite executing git commands via os/exec. All other
git execution in the codebase lives in internal/workspace. Move the three
functions to a new internal/workspace/overlaysync.go, keep DeriveOverlayURL and
OverlayDir in config (pure URL/path computation), and update callers in
internal/cli/init.go to call workspace.CloneOrSyncOverlay and workspace.HeadSHA.
